+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added base price (purchase price) support
 - Added base price converting
 - Added customer order status new
+- Added customer / customer order shipping / billing department support
 - Changed pull value handling - stripping whitespaces from the beginning and end of a string
 - Changed image save and delete error handling
 - Fixed payment price value to gross